requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
lxml>=5.2.1